complete cusotm hours and schedule service tests

+++ SCS-Backend/src/main/java/ca/mcgill/ecse321/scs/service/CustomHoursService.java
@@ -7,10 +7,12 @@ import java.time.LocalTime;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import ca.mcgill.ecse321.scs.dao.CustomHoursRepository;
+import ca.mcgill.ecse321.scs.exception.SCSException;
 import ca.mcgill.ecse321.scs.model.CustomHours;
 
 import ca.mcgill.ecse321.scs.model.Schedule;
@@ -22,8 +24,25 @@ public class CustomHoursService {
     @Autowired
     ScheduleService scheduleService;
 
+    public void setScheduleService(ScheduleService scheduleService) {
+        this.scheduleService = scheduleService;
+    }
+    
     @Transactional
     public CustomHours createCustomHours(String name, String description, LocalDate date, LocalTime openTime, LocalTime closeTime, int year) {
+        // if closeTime before openTime, throw exception
+        if (closeTime.isBefore(openTime)) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Close time cannot be before open time.");
+        } else if (customHoursRepository.findCustomHoursByName(name) != null) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Custom hours with name " + name + " already exists.");
+        } else if (name == null || name.trim().length() == 0) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Name cannot be empty.");
+        } else if (description == null || description.trim().length() == 0) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Description cannot be empty.");
+        } else if (year != date.getYear()) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Year of date does not match year of schedule.");
+        }
+
         CustomHours customHours = new CustomHours();
         customHours.setName(name);
         customHours.setDescription(description);
@@ -39,7 +58,39 @@ public class CustomHoursService {
     }
 
     @Transactional
+    public CustomHours getCustomHours(String name) {
+        CustomHours customHours = customHoursRepository.findCustomHoursByName(name);
+        if (customHours == null) {
+            throw new SCSException(HttpStatus.NOT_FOUND, "Custom hours with name " + name + " does not exist.");
+        }
+        return customHours;
+    }
+
+    @Transactional
+    public CustomHours getCustomHoursByDate(LocalDate date) {
+        // there can only be 1 custom hours for a given date
+        List<CustomHours> customHours = ServiceUtils.toList(customHoursRepository.findAll());
+        for (CustomHours ch : customHours) {
+            if (ch.getDate().equals(Date.valueOf(date))) {
+                return ch;
+            }
+        }
+        
+        throw new SCSException(HttpStatus.NOT_FOUND, "Custom hours for date " + date + " does not exist.");
+    }
+
+    @Transactional
     public CustomHours updateCustomHours(String name, String description, LocalDate date, LocalTime openTime, LocalTime closeTime, int year) {
+        if (closeTime.isBefore(openTime)) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Close time cannot be before open time.");
+        } else if (name == null || name.trim().length() == 0) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Name cannot be empty.");
+        } else if (description == null || description.trim().length() == 0) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Description cannot be empty.");
+        } else if (year != date.getYear()) {
+            throw new SCSException(HttpStatus.BAD_REQUEST, "Year of date does not match year of schedule.");
+        }
+
         CustomHours customHours = customHoursRepository.findCustomHoursByName(name);
         customHours.setDescription(description);
         customHours.setDate(Date.valueOf(date));
@@ -54,24 +105,6 @@ public class CustomHoursService {
     }
 
     @Transactional
-    public CustomHours getCustomHours(String name) {
-        return customHoursRepository.findCustomHoursByName(name);
-    }
-
-    @Transactional
-    public CustomHours getCustomHoursByDate(LocalDate date) {
-        // there can only be 1 custom hours for a given date
-        List<CustomHours> customHours = ServiceUtils.toList(customHoursRepository.findAll());
-        for (CustomHours ch : customHours) {
-            if (ch.getDate().equals(Date.valueOf(date))) {
-                return ch;
-            }
-        }
-        
-        return null;
-    }
-
-    @Transactional
     public List<CustomHours> getAllCustomHours() {
         return ServiceUtils.toList(customHoursRepository.findAll());
     }
@@ -79,6 +112,9 @@ public class CustomHoursService {
     @Transactional
     public void deleteCustomHours(String name) {
         CustomHours customHours = customHoursRepository.findCustomHoursByName(name);
+        if (customHours == null) {
+            throw new SCSException(HttpStatus.NOT_FOUND, "Custom hours with name " + name + " not found.");
+        }
         customHoursRepository.delete(customHours);
     }
